% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Coords.R
\name{Coords}
\alias{Coords}
\alias{[.Coords}
\alias{dim.Coords}
\alias{+.Coords}
\alias{add.Coords}
\alias{-.Coords}
\alias{subtract.Coords}
\alias{*.Coords}
\alias{scale.Coords}
\title{Coordinates class}
\usage{
\method{[}{Coords}(x, i)

\method{dim}{Coords}(x)

\method{+}{Coords}(a, b, na.rm = FALSE)

add.Coords(a, b, na.rm = FALSE)

\method{-}{Coords}(a, b, na.rm = FALSE)

subtract.Coords(a, b, na.rm = FALSE)

\method{*}{Coords}(x, k)

\method{scale}{Coords}(x, k)
}
\description{
A class to manage the basic functions of an n dimension coordinate system.
}
\section{Functions}{
\itemize{
\item \code{[.Coords}: Subset by index as in any numeric vector x\link{i}.

\item \code{dim.Coords}: Number of dimension of the coordinate system.

\item \code{+.Coords}: Sum \code{Coords} objects as in \code{A + B}.

\item \code{add.Coords}: Add two \code{Coords} objects with \code{add(A, B)}.

\item \code{-.Coords}: Subtract \code{Coords} objects as in \code{A - B}.

\item \code{subtract.Coords}: Subtract two \code{Coords} objects with \code{subtract(A, B)}.

\item \code{*.Coords}: Scale a \code{Coords} object with \code{A * k}.

\item \code{scale.Coords}: Scale a \code{Coords} object with \code{scale(A, k)}.
}}

\references{
\link{https://www.redblobgames.com/grids/hexagons}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{Numeric, the value of the coordinates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Coords$new()}}
\item \href{#method-set}{\code{Coords$set()}}
\item \href{#method-get}{\code{Coords$get()}}
\item \href{#method-dim}{\code{Coords$dim()}}
\item \href{#method-print}{\code{Coords$print()}}
\item \href{#method-clone}{\code{Coords$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{Coords}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the numeric value of the coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$set(x, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get the numeric value of the coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$get(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{integer, the index(es) to be retrieved. If missing, the full set of
coordinates is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
Get the number of dimensions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\code{print} method for class \code{Coords}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
