% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-CubeCoords.R
\name{CubeCoords}
\alias{CubeCoords}
\title{Cube coordinates class}
\description{
A class \code{Coords}of cube coordinates.
}
\section{Super class}{
\code{\link[hexagon:Coords]{hexagon::Coords}} -> \code{CubeCoords}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{q}}{numeric, the 'q' dimension.}

\item{\code{r}}{numeric, the 'r' dimension.}

\item{\code{s}}{numeric, the 's' dimension.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CubeCoords$new()}}
\item \href{#method-set}{\code{CubeCoords$set()}}
\item \href{#method-get}{\code{CubeCoords$get()}}
\item \href{#method-clone}{\code{CubeCoords$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Coords" data-id="dim">}\href{../../hexagon/html/Coords.html#method-dim}{\code{hexagon::Coords$dim()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Coords" data-id="print">}\href{../../hexagon/html/Coords.html#method-print}{\code{hexagon::Coords$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeCoords$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric coordinates of length 3.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{CubeCoords}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the coordinates value.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeCoords$set(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Numeric coordinates of length 3.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Invisible self.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get coodinate values.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeCoords$get(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{The index of the dimension to be retrieved or
the name as a string 'q', 'r' or 's'.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric value. As many as \code{NA_real_} as there are indexes
out of boundaries.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CubeCoords$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
