% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-HexCubeCenter.R
\name{direction}
\alias{direction}
\alias{hex_direction,hex_diagonal_direction}
\alias{hex_direction}
\alias{hex_diagonal_direction}
\title{Cube direction vectors from an hexagon}
\usage{
hex_direction(direction)

hex_diagonal_direction(direction)
}
\arguments{
\item{direction}{a number between 1 and 6.}
}
\value{
A numeric vector of length 3.
}
\description{
\code{hex_direction} and \code{hex_diagonal_direction} return a vector
towards one of the the six possible directions in the hexagon's neighbour.
They are ued in the neighbour methods \code{hex_neighbour} and
\code{hex_diagonal_neighbour}. Given two \code{HexCubeCenter} A and B, where A is a
generic center and B is the center build on the output of the direction
functions, the sum A + B gives the coordinates of neighbour hexagon in the
specified direction.
}
\section{Functions}{
\itemize{
\item \code{hex_direction}: Directions of the immediate neighbour.

\item \code{hex_diagonal_direction}: Directions of the diagonal neighbour.
}}

\references{
\link{https://www.redblobgames.com/grids/hexagons}
}
