% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_concetric.R
\name{.coords_grid_concetric}
\alias{.coords_grid_concetric}
\title{Coordinates of a concentric hex grid}
\usage{
.coords_grid_concetric(n)
}
\arguments{
\item{n}{the number of rings around an origin hexagon. It determines the
extent of the grid.}
}
\value{
A list of cube coordinates.
}
\description{
Compute the coordinates of a concentric hex grid of \code{n} rings.
}
