% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_concetric.R
\name{.hex_cube_center_v}
\alias{.hex_cube_center_v}
\title{Cube centers of a concentric hex grid}
\usage{
.hex_cube_center_v(coords_list)
}
\arguments{
\item{coords_list}{A list of vectors, each containing cube coordinates.}
}
\value{
A list of objects of class \code{HexCubeCenter}.
}
\description{
Instantiate the \code{HexCubeCenter} of a concentric hex grid of \code{n} rings.
}
