% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_names}
\alias{generate_names}
\title{Random names for your hexagons}
\usage{
generate_names(n)
}
\arguments{
\item{n}{How many names to generate}
}
\description{
Generate random couples of adjectives and nouns. It can be used
to generate names for your hexagon in a grid.
}
