% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Center.R
\name{hex_diagonal_neighbor.HexCubeCenter}
\alias{hex_diagonal_neighbor.HexCubeCenter}
\title{Hexagon diagonal neighbour}
\usage{
\method{hex_diagonal_neighbor}{HexCubeCenter}(hex, direction)
}
\value{
An object of class \code{HexCubeCenter}.
}
\description{
\code{hex_diagonal_neighbour} method for class HexCubeCenter
}
