% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_concetric.R
\name{hex_grid_concentric}
\alias{hex_grid_concentric}
\title{Concentric Hex Grid}
\usage{
hex_grid_concentric(
  hexagon_radius = 100,
  n = 5,
  planar_origin = NULL,
  crs_input = 3857,
  crs_output = 4326,
  pointy_top = TRUE
)
}
\arguments{
\item{hexagon_radius}{the radius of the hexagon's circumscribed circle.}

\item{n}{the number of rings around the origin hexagon.}

\item{planar_origin}{A vector of geographical coordinates.
If \code{NULL}, it is set to 0,0.}

\item{crs_input}{A number. It determines the geographical reference
of the input.}

\item{crs_output}{A number. It determines the geographical reference
of the output.}

\item{pointy_top}{Logical.}

\item{identifier_name}{character, the name of the identifier column
of the output.}
}
\value{
An \code{sfg} object.
}
\description{
Create a concentric hex grid in the geographical space. Concetric
grids are built in concentric rings around an hexagon origin. The overall
shape resembles an hexagon and the size is roughly \code{4 * n * hexagon_radius}
}
\references{
\link{https://www.redblobgames.com/grids/hexagons}
}
