% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-HexCubeCenter.R
\name{hex_neighbour.HexCubeCenter}
\alias{hex_neighbour.HexCubeCenter}
\title{Hexagon neighbour}
\usage{
\method{hex_neighbour}{HexCubeCenter}(hex, direction)
}
\value{
An object of class \code{HexCubeCenter}.
}
\description{
\code{hex_neighbour} method for class HexCubeCenter
}
