% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_concetric.R
\name{hex_ring_grid}
\alias{hex_ring_grid}
\alias{hex_ring_grid_wkt}
\title{Concentric Hex Grid}
\usage{
hex_ring_grid(
  hexagon_radius = 100,
  n = 5,
  planar_origin = NULL,
  crs_input = 3857,
  crs_output = 4326,
  pointy_top = TRUE
)

hex_ring_grid_wkt(
  hexagon_radius = 100,
  n = 5,
  planar_origin = NULL,
  crs_input = 3857,
  crs_output = 4326,
  pointy_top = TRUE
)
}
\arguments{
\item{hexagon_radius}{the radius of the hexagon's circumscribed circle.}

\item{n}{the number of rings around the origin hexagon.}

\item{planar_origin}{A vector of geographical coordinates.
If \code{NULL}, it is set to 0,0.}

\item{crs_input}{A number. It determines the geographical reference
of the input.}

\item{crs_output}{A number. It determines the geographical reference
of the output.}

\item{pointy_top}{Logical.}

\item{identifier_name}{character, the name of the identifier column
of the output.}
}
\value{
An \code{sfg} object.
}
\description{
Create a concentric hex grid in the geographical space. Concetric
grids are built in concentric rings around an hexagon origin. The overall
shape resembles an hexagon and the size is roughly \code{4 * n * hexagon_radius}
}
\section{Functions}{
\itemize{
\item \code{hex_ring_grid_wkt}: Return a character vector of WKT geometries.
}}

\references{
\link{https://www.redblobgames.com/grids/hexagons}
}
