% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_concetric.R
\name{hex_ring_grid_count}
\alias{hex_ring_grid_count}
\title{Count of element in a concentric grid}
\usage{
hex_ring_grid_count(n = 5)
}
\arguments{
\item{n}{the number of rings around the origin hexagon.}
}
\description{
Count of element in a concentric grid
}
