% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hexagon.R
\name{Hexagon}
\alias{Hexagon}
\title{Hexagon classes}
\description{
Hexagons are the primitive unit to build hex grids. They can be pointy
or flat, depending on whether the top is a vertex or an edge.

An hexagon is defined by its size and its position within the grid. A geographical
hexagon is also defined by its geographical position.
}
\section{Geographical reference}{


In order to place the hexagon in place, a geographical origin must be
specified. The geographical origin is assumed to be in planar coordinates.
}

\references{
\link{https://www.redblobgames.com/grids/hexagons}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{center}}{\code{CubeCoords} object.}

\item{\code{size}}{numeric. The cartesian distance between
the center and any vertex of the hexagon.}

\item{\code{planar_origin}}{numeric of length 2. The planar coordinates of the origin.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Hexagon$new()}}
\item \href{#method-deep_clone}{\code{Hexagon$deep_clone()}}
\item \href{#method-v_planar_coordinates}{\code{Hexagon$v_planar_coordinates()}}
\item \href{#method-is_pointy}{\code{Hexagon$is_pointy()}}
\item \href{#method-get_planar_center}{\code{Hexagon$get_planar_center()}}
\item \href{#method-set_center}{\code{Hexagon$set_center()}}
\item \href{#method-set_planar_origin}{\code{Hexagon$set_planar_origin()}}
\item \href{#method-set_size}{\code{Hexagon$set_size()}}
\item \href{#method-to_polygon}{\code{Hexagon$to_polygon()}}
\item \href{#method-clone}{\code{Hexagon$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$new(center, size = NULL, planar_origin = NULL, pointy_top = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{A vector of cube coordinates or an object that can be coerced to
a \code{HexCubeCenter}.}

\item{\code{size}}{A number defining the radius of the circumscribed circle
of the hexagon, in the units of the planar_origin.}

\item{\code{planar_origin}}{A vector of geographical coordinates.
If \code{NULL}, it is set to 0,0.}

\item{\code{pointy_top}}{Logical.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-deep_clone"></a>}}
\if{latex}{\out{\hypertarget{method-deep_clone}{}}}
\subsection{Method \code{deep_clone()}}{
Deep clone function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$deep_clone()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-v_planar_coordinates"></a>}}
\if{latex}{\out{\hypertarget{method-v_planar_coordinates}{}}}
\subsection{Method \code{v_planar_coordinates()}}{
Compute vertices' geographical coordinates in
planar coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$v_planar_coordinates()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A numeric matrix 6 by 2.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-is_pointy"></a>}}
\if{latex}{\out{\hypertarget{method-is_pointy}{}}}
\subsection{Method \code{is_pointy()}}{
Whether the hexagon is pointy or flat.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$is_pointy()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_planar_center"></a>}}
\if{latex}{\out{\hypertarget{method-get_planar_center}{}}}
\subsection{Method \code{get_planar_center()}}{
Compute the geographical coordinates of the hex
center in planar coordinate system.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$get_planar_center()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \code{Coords}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_center"></a>}}
\if{latex}{\out{\hypertarget{method-set_center}{}}}
\subsection{Method \code{set_center()}}{
Set the center of the hexagon.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$set_center(x, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A set of cube coordinates or something that can be coerced to it.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_planar_origin"></a>}}
\if{latex}{\out{\hypertarget{method-set_planar_origin}{}}}
\subsection{Method \code{set_planar_origin()}}{
Set the the planar origin.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$set_planar_origin(x, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{Geographical coordinates.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_size"></a>}}
\if{latex}{\out{\hypertarget{method-set_size}{}}}
\subsection{Method \code{set_size()}}{
Set the size of the hexagon. If \code{NULL} the size is set to 1.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$set_size(size = NULL, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{size}}{A number defining the radius of the circumscribed circle
of the hexagon, in the units of the planar_origin.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_polygon"></a>}}
\if{latex}{\out{\hypertarget{method-to_polygon}{}}}
\subsection{Method \code{to_polygon()}}{
Compute the geographical
representation of the polygon in planar coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$to_polygon()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
An object of class \verb{XY, POLYGON, sfg} (see package \code{sf}).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Hexagon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
