% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{initiate_instance}
\alias{initiate_instance}
\title{Initiate instance}
\usage{
initiate_instance(..., what)
}
\arguments{
\item{...}{Named arguments of the \verb{$new} method of the instance to be
instantiated.}

\item{what}{character, the name of the class to be instantiated.}
}
\value{
An object of class as specified in \code{what}.
}
\description{
Programatically initiate an R6 class.
}
