% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Hexagon.R
\name{PointyHexagon}
\alias{PointyHexagon}
\title{Pointy Hexagon class}
\description{
A class \code{Hexagon} with a pointy top.
}
\section{Super class}{
\code{\link[hexagon:Hexagon]{hexagon::Hexagon}} -> \code{PointyHexagon}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{width}}{the distance between two edges of the hexagon in
the units of \code{size}.}

\item{\code{height}}{the distance between two vertices of the hexagon in
the units of \code{size}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{PointyHexagon$new()}}
\item \href{#method-shift_on_q}{\code{PointyHexagon$shift_on_q()}}
\item \href{#method-shift_on_r}{\code{PointyHexagon$shift_on_r()}}
\item \href{#method-shift_on_s}{\code{PointyHexagon$shift_on_s()}}
\item \href{#method-shift_along_direction}{\code{PointyHexagon$shift_along_direction()}}
\item \href{#method-shift_along_diagonal_direction}{\code{PointyHexagon$shift_along_diagonal_direction()}}
\item \href{#method-clone}{\code{PointyHexagon$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="deep_clone">}\href{../../hexagon/html/Hexagon.html#method-deep_clone}{\code{hexagon::Hexagon$deep_clone()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="get_planar_center">}\href{../../hexagon/html/Hexagon.html#method-get_planar_center}{\code{hexagon::Hexagon$get_planar_center()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="is_pointy">}\href{../../hexagon/html/Hexagon.html#method-is_pointy}{\code{hexagon::Hexagon$is_pointy()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="set_center">}\href{../../hexagon/html/Hexagon.html#method-set_center}{\code{hexagon::Hexagon$set_center()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="set_crs_input">}\href{../../hexagon/html/Hexagon.html#method-set_crs_input}{\code{hexagon::Hexagon$set_crs_input()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="set_crs_output">}\href{../../hexagon/html/Hexagon.html#method-set_crs_output}{\code{hexagon::Hexagon$set_crs_output()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="set_planar_origin">}\href{../../hexagon/html/Hexagon.html#method-set_planar_origin}{\code{hexagon::Hexagon$set_planar_origin()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="set_size">}\href{../../hexagon/html/Hexagon.html#method-set_size}{\code{hexagon::Hexagon$set_size()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="to_polygon">}\href{../../hexagon/html/Hexagon.html#method-to_polygon}{\code{hexagon::Hexagon$to_polygon()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hexagon" data-topic="Hexagon" data-id="v_planar_coordinates">}\href{../../hexagon/html/Hexagon.html#method-v_planar_coordinates}{\code{hexagon::Hexagon$v_planar_coordinates()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$new(
  center,
  size = NULL,
  planar_origin = NULL,
  crs_input = 3857,
  crs_output = crs_input
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{A vector of cube coordinates or an object that can be coerced to
a \code{HexCubeCenter}.}

\item{\code{size}}{A number defining the radius of the circumscribed circle
of the hexagon, in the units of the planar_origin.}

\item{\code{planar_origin}}{A vector of geographical coordinates.
If \code{NULL}, it is set to 0,0.}

\item{\code{crs_input}}{A number. It determines the geographical reference
of the input.}

\item{\code{crs_output}}{A number. It determines the geographical reference
of the output.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{PointyHexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_on_q"></a>}}
\if{latex}{\out{\hypertarget{method-shift_on_q}{}}}
\subsection{Method \code{shift_on_q()}}{
Shift the hex center of \code{i} places in the cube system,
along the \code{q} axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$shift_on_q(i, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the distance of the shift in number of cells.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In place or a new \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_on_r"></a>}}
\if{latex}{\out{\hypertarget{method-shift_on_r}{}}}
\subsection{Method \code{shift_on_r()}}{
Shift the hex center of \code{i} places in the cube system,
along the \code{r} axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$shift_on_r(i, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the distance of the shift in number of cells.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In place or a new \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_on_s"></a>}}
\if{latex}{\out{\hypertarget{method-shift_on_s}{}}}
\subsection{Method \code{shift_on_s()}}{
Shift the hex center of \code{i} places in the cube system,
along the \code{s} axis.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$shift_on_s(i, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the distance of the shift in number of cells.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In place or a new \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_along_direction"></a>}}
\if{latex}{\out{\hypertarget{method-shift_along_direction}{}}}
\subsection{Method \code{shift_along_direction()}}{
Shift along a direction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$shift_along_direction(i, direction, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the distance of the shift in number of cells.}

\item{\code{direction}}{a number from 1 to 6, corresponding to a direction. 1
in a pointy hexagon is to the right and the other are in sequence
anti-clockwise.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In place or a new \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shift_along_diagonal_direction"></a>}}
\if{latex}{\out{\hypertarget{method-shift_along_diagonal_direction}{}}}
\subsection{Method \code{shift_along_diagonal_direction()}}{
Shift along a diagonal direction.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$shift_along_diagonal_direction(i, direction, in_place = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{the distance of the shift in number of cells.}

\item{\code{direction}}{a number from 1 to 6, corresponding to a direction.}

\item{\code{in_place}}{logical. If TRUE the instance is modifed by reference and
cloned otherwise.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
In place or a new \code{Hexagon}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PointyHexagon$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
