% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switchv}
\alias{switchv}
\title{A vectorised switch function}
\usage{
switchv(EXPR, ...)
}
\arguments{
\item{EXPR}{an expression evaluating to a number or a character
    string.}

\item{...}{the list of alternatives.  If it is intended that
    \code{EXPR} has a character-string value these will be
    named, perhaps except for one alternative to be used as a
    \sQuote{default} value.}
}
\value{
A vector of elements of \code{...}.
}
\description{
switchv evaluates a vector of EXPR and accordingly chooses one of the
further arguments (in ...).
}
