% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_hexagon}
\alias{theme_hexagon}
\title{theme_hexagon}
\usage{
theme_hexagon(hexdata)
}
\arguments{
\item{base_size}{Base font size, defaults to 11}

\item{base_family}{Font family.}

\item{dark}{Primary forefront color.}

\item{light}{Primary background color.}

\item{gray}{Gray color.}
}
\value{
My ggplot2 theme (similar to theme_minimal/theme_bw)
}
\description{
A hex sticker friendly ggplot2 theme
}
