% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Coords.R
\name{Coords}
\alias{Coords}
\title{Coordinates class}
\description{
A class to manage the basic functions of the coordinates.
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{value}}{Numeric, the value of the coordinates.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Coords$new()}}
\item \href{#method-set}{\code{Coords$set()}}
\item \href{#method-get}{\code{Coords$get()}}
\item \href{#method-dim}{\code{Coords$dim()}}
\item \href{#method-print}{\code{Coords$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialise a new instance.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An object of class \code{Coords}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set"></a>}}
\if{latex}{\out{\hypertarget{method-set}{}}}
\subsection{Method \code{set()}}{
Set the numeric value of the coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$set(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{A numeric vector.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get"></a>}}
\if{latex}{\out{\hypertarget{method-get}{}}}
\subsection{Method \code{get()}}{
Get the numeric value of the coordinates.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$get(i)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{i}}{integer, the index(es) to be retrieved. If missing, the full set of
coordinates is returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A numeric vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dim"></a>}}
\if{latex}{\out{\hypertarget{method-dim}{}}}
\subsection{Method \code{dim()}}{
Get the number of dimensions.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$dim()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A number.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\code{print} method for class \code{Coords}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Coords$print()}\if{html}{\out{</div>}}
}

}
}
